pyproject.toml
[project]
name = "pereobuyka"
version = "0.1.0"
description = "Telegram-бот-консультант для записи в сервис шиномонтажа."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.0.0",
  "openai>=1.0.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
pereobuyka = "pereobuyka.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

